requirements.txt
# --- Core API ---
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-multipart

# --- Base de Datos ---
sqlalchemy
psycopg2-binary

# --- Mensajería (RabbitMQ) ---
aio-pika

# --- Seguridad & Cliente HTTP (Keycloak/Webhooks) ---
python-jose[cryptography]
httpx

# --- Observabilidad (OpenTelemetry) ---
# Estas son las nuevas para conectar con Jaeger y Prometheus
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp
prometheus-client
opentelemetry-exporter-otlp-proto-http
python-json-logger==2.0.7